+++ app/controllers/application_controller.rb
@@ -1,7 +1,5 @@
 class ApplicationController < ActionController::Base
-    helper_method :current_user
-    before_action :authenticate_user!
-    
+    # helper_method :current_user
     # before_action :login_required
     
     # def login_required
