+++ app/controllers/users/application_controller.rb
@@ -1,0 +1,5 @@
+module Users
+  class ApplicationController < ::ApplicationController
+    before_action :authenticate_user!
+  end
+end
