+++ app/views/users/posts/new.html.slim
@@ -1,0 +1,7 @@
+h1 投稿画面
+.nav.d-flex.justify-content-end
+  = link_to'一覧',posts_path, class: 'nav-link'
+
+
+/ ボタン部分_form.html.slimから呼び出し
+= render partial:'form',locals:{ post: @post}
