+++ config/routes.rb
@@ -1,6 +1,9 @@
 Rails.application.routes.draw do
   devise_for :users, controllers: { omniauth_callbacks: 'users/omniauth_callbacks' }
   root to: 'posts#index'
-  resources :posts
+  resources :posts, only: %w[index show]
+  namespace :users do
+    resources :posts
+  end
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
